+++ app/controllers/shoppe/products_controller.rb
@@ -5,7 +5,7 @@ module Shoppe
     before_filter { params[:id] && @product = Shoppe::Product.find(params[:id]) }
   
     def index
-      @products = Shoppe::Product.includes(:stock_level_adjustments, :default_image, :product_category).order(:title).group_by(&:product_category).sort_by { |cat,pro| cat.name }
+      @products = Shoppe::Product.includes(:stock_level_adjustments, :default_image, :product_category).order(:name).group_by(&:product_category).sort_by { |cat,pro| cat.name }
     end
   
     def new
@@ -54,7 +54,7 @@ module Shoppe
     private
   
     def safe_params
-      params[:product].permit(:product_category_id, :title, :sku, :permalink, :description, :short_description, :weight, :price, :cost_price, :tax_rate_id, :stock_control, :default_image_file, :data_sheet_file, :active, :featured, :in_the_box, :product_attributes_array => [:key, :value, :searchable, :public])
+      params[:product].permit(:product_category_id, :name, :sku, :permalink, :description, :short_description, :weight, :price, :cost_price, :tax_rate_id, :stock_control, :default_image_file, :data_sheet_file, :active, :featured, :in_the_box, :product_attributes_array => [:key, :value, :searchable, :public])
     end
   
   end