+++ db/migrate/20131022092904_rename_product_title_to_name.rb
@@ -1,0 +1,5 @@
+class RenameProductTitleToName < ActiveRecord::Migration
+  def change
+    rename_column :shoppe_products, :title, :name
+  end
+end